- add new interface CarService.java and in this interface add new contract findAllCars(); - changed name a java class from CarService -> CarServiceImp;
